requirements.txt
python-telegram-bot==20.7
openai==1.97.0
pdfminer.six==20250506
python-docx==1.2.0
docx2txt==0.9
pytesseract==0.3.13
Pillow==11.3.0
fpdf2==2.8.3

